predict合并前的版本

+++ include/rviz_draw.h
@@ -11,7 +11,7 @@
 namespace rviz_draw
 {
     visualization_msgs::Marker draw(
-        std::vector<Eigen::Vector3f> points,
+        std::vector<Eigen::Vector3d> points,
         float* colar,
         const char* frameId="temp", 
         int id=1,
